% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_check.R
\name{auth_checks}
\alias{auth_checks}
\title{Check the username and password given in input$user and input$password respectivly in the current namespace
If the credentuals are correct return the user information}
\usage{
auth_checks(input, output, session, con, user_id_col_name, password_col_name,
  user_tabel)
}
\description{
Check the username and password given in input$user and input$password respectivly in the current namespace
If the credentuals are correct return the user information
}
