% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_server.R
\name{auth_server}
\alias{auth_server}
\title{Manage the logon, and the changes of the UI based upon}
\usage{
auth_server(con, logged_on_server, user_id_col_name, password_col_name,
  user_table)
}
\arguments{
\item{con}{Database connection or pool}
}
\description{
Manage the logon, and the changes of the UI based upon
}
