% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_sidebar.R
\name{auth_sidebar}
\alias{auth_sidebar}
\title{Create the sidbar for the dashboard, diffrent dependent upon the status of the app}
\usage{
auth_sidebar(input, output, session, status)
}
\arguments{
\item{status}{What is the status of the app and as such which of the sidebars should be shown}

\item{tabs}{Which of the tabs should be shown, if left it defults to NULL showing all possible tabs}

\item{competencies}{The list containg detals of all of the possible tabs}
}
\description{
Create the sidbar for the dashboard, diffrent dependent upon the status of the app
}
