% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_functions.R
\name{fixed_shiny_input}
\alias{fixed_shiny_input}
\title{Defult syling for what was a slect input that another user has changed and saved}
\usage{
fixed_shiny_input(label = NULL, text = NULL)
}
\description{
Defult syling for what was a slect input that another user has changed and saved
}

