% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_password.R
\name{save_new_password}
\alias{save_new_password}
\title{Save a new password to the db}
\usage{
save_new_password(session, con, user_id_col_name, password_col_name, user_tabe,
  user_id, new_hased_password)
}
\description{
Save a new password to the db
}
