% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_tab.R
\name{save_user_details}
\alias{save_user_details}
\title{Save the current user datailes to the db}
\usage{
save_user_details(input, output, session, auth, changeable_columns)
}
\description{
Save the current user datailes to the db
}
