% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_tab.R
\name{settings_tab}
\alias{settings_tab}
\title{Create the settings tab
This function needs to be able to trigger the fech of a new dt_user}
\usage{
settings_tab(input, output, session, auth, dt_settings_user = NULL)
}
\arguments{
\item{dt_settings_user}{Data table for the user for which to show settings
only valid on the admin or lm page}

\item{dt_user}{Data table identifying the current user}
}
\description{
Create the settings tab
This function needs to be able to trigger the fech of a new dt_user
}
